+++ aiida/orm/implementation/calculation.py
@@ -21,14 +21,10 @@ from aiida.orm.implementation.general.calculation.job import _input_subfolder
 if BACKEND == BACKEND_SQLA:
     from aiida.orm.implementation.sqlalchemy.calculation import Calculation
     from aiida.orm.implementation.sqlalchemy.calculation.job import JobCalculation
-    from aiida.orm.implementation.sqlalchemy.calculation.inline import (
-        InlineCalculation, make_inline)
 
 elif BACKEND == BACKEND_DJANGO:
     from aiida.orm.implementation.django.calculation import Calculation
     from aiida.orm.implementation.django.calculation.job import JobCalculation
-    from aiida.orm.implementation.django.calculation.inline import (
-        InlineCalculation, make_inline)
 
 else:
     raise ConfigurationError("Invalid settings.BACKEND: {}".format(
